:recycle: smartened up the code + discretised x, y
