--- analysis/face_model.py
@@ -1,96 +1,0 @@
-from matplotlib import pyplot as plt
-import gpytorch
-import torch
-
-
-
-
-
-class PlaneModel():
-    def __init__(self, pos1, pos2, pos3):
-        # A plane is defined by Ax + By + CT = D where A, B, C are the components of the normals and D is a constant
-        # Hence to find the T at a specific position in the plane all we need is T = (D - Ax - By)/C
-        normal = torch.cross(pos2 - pos1, pos3 - pos1)
-        self.__A = normal[0]
-        self.__B = normal[1]
-        self.__C = normal[2]
-        self.__D = torch.dot(normal, pos1)
-    
-
-    def get_T(self, pos_xy):
-        # We return the temperature at a specific x and y position
-        temperature_tensor = (self.__D - self.__A * pos_xy[:, 0] - self.__B * pos_xy[:, 1])/self.__C
-        return torch.reshape(temperature_tensor, (-1, 1))
-
-
-
-
-
-
-class ExactGP(gpytorch.models.ExactGP):
-    def __init__(self, x_train, y_train, likelihood):
-        super().__init__(x_train, y_train, likelihood)
-
-        self.mean_module = gpytorch.means.ConstantMean()
-        self.covar_module = gpytorch.kernels.ScaleKernel(gpytorch.kernels.RBFKernel())
-        self.likelihood = likelihood
-
-        self.likelihood.noise = 1e-3
-        all_params = set(self.parameters())
-        final_params = list(all_params - {self.likelihood.raw_noise})
-        self.optimiser = torch.optim.Adam(final_params, lr=0.1)
-
-
-    def forward(self, x):
-        # Feed forward the input x
-        mean_x = self.mean_module(x)
-        covar_x = self.covar_module(x)
-        return gpytorch.distributions.MultivariateNormal(mean_x, covar_x)
-
-
-
-
-
-class GaussianManager():
-    def __init__(self, positions):
-        train_xy = positions[:, :2]
-        train_T = positions[:, 2]
-
-        self.__likelihood = gpytorch.likelihoods.GaussianLikelihood()
-        self.__model = ExactGP(train_xy, train_T, self.__likelihood)
-        self.setup_model(train_xy, train_T)
-        
-
-    def setup_model(self, train_xy, train_T):
-        self.__model.train()
-        self.__likelihood.train()
-        optimizer = torch.optim.Adam(self.__model.parameters(), lr=0.1)
-        mll = gpytorch.mlls.ExactMarginalLogLikelihood(self.__likelihood, self.__model)
-        
-        for i in range(50):
-            optimizer.zero_grad()
-            # Output from model
-            output = self.__model(train_xy)
-            # Calc loss and backprop gradients
-            loss = -mll(output, train_T)
-            loss.backward(retain_graph=True)
-            optimizer.step()
-
-
-    def get_T(self, pos):
-        self.__model.eval()
-        return self.__model(pos).mean
-
-
-
-if __name__ == "__main__":
-    # plane = PlaneModel(torch.tensor([1, 2, 3]), torch.tensor([2, 3, 4]), torch.tensor([2, 4, 5]))
-    # print(plane.get_T(torch.tensor([[1, 2], [2, 3], [2, 4]])))
-
-    positions = torch.Tensor([
-        [1, 2, 3],
-        [2, 3, 4],
-        [2, 4, 5]
-    ])
-    manager = GaussianManager(positions)
-    print(manager.get_T(torch.Tensor([[1, 2], [2, 3], [2, 5]])))
