+++ src/optimiser_manager.py
@@ -1,0 +1,16 @@
+from matplotlib import pyplot as plt
+from csv_reader import CSVReader
+from tqdm import tqdm
+import numpy as np
+
+
+def f(pos):
+    return np.sum(np.sin(pos))
+
+
+
+
+
+
+if __name__ == "__main__":
+    pass
